move TestDebugLogger to the unittests
